expriments

+++ contrib/sqms/src/include/common/config.h
@@ -35,7 +35,7 @@ extern long defaultTTLSeconds;
 extern bool debug;
 
 extern const int max_msg_size;
-extern const int shared_mem_size;
+extern const size_t shared_mem_size;
 extern const char* queue_name;
 extern const char* shared_index_name;
